Rename "generalUtils"
