+++ app/meetings/pages/meeting/create.tsx
@@ -7,7 +7,7 @@ import { Button, Modal } from "react-bootstrap"
 import Card from "react-bootstrap/Card"
 import { CopyToClipboard } from "react-copy-to-clipboard"
 import Skeleton from "react-loading-skeleton"
-import { getOrigin } from "utils/generalUtils"
+import { getOrigin } from "utils/origin"
 import * as z from "zod"
 import General from "../../components/creationSteps/General"
 import ScheduleStep from "../../components/creationSteps/Schedule"
