+++ app/meetings/utils/format-meeting-href.ts
@@ -1,5 +1,5 @@
 import type { Meeting } from "db"
-import { getOrigin } from "utils/generalUtils"
+import { getOrigin } from "utils/origin"
 
 export function formatMeetingHref(meeting: Pick<Meeting, "ownerName" | "link">) {
   return `${getOrigin()}/schedule/${meeting.ownerName}/${meeting.link}`
